Create mixin to replace duplicate code
