+++ lib/vast/element.rb
