+++ lib/vast/non_linear_creative.rb
@@ -1,7 +1,15 @@
+require_relative 'resource.rb'
+
 module VAST
   # NonLinearCreative runs concurrently with the video content so the users see the ad while viewing 
   # the content. Non-linear video ads can be delivered as text, graphical ads, or as video overlays.
   class NonLinearCreative < Creative
+    include VAST::Resource
+
+    def initialize(node)
+      super(node)
+      after_initialize(node)
+    end
 
     def id
       source_node[:id]
@@ -56,42 +64,5 @@ module VAST
     def min_suggested_duration
       source_node[:minSuggestedDuration]
     end
-    
-    
-    # Type of non-linear resource, returned as a symbol. Either :static, :iframe, or :html.
-    def resource_type
-      if source_node.at('StaticResource')
-        :static
-      elsif source_node.at('IFrameResource')
-        :iframe
-      elsif source_node.at('HTMLResource')
-        :html
-      end
-    end
-    
-    # Returns MIME type of static creative
-    def creative_type
-      if resource_type == :static
-        source_node.at('StaticResource')[:creativeType]
-      end
-    end
-    
-    # Returns URI for static or iframe resource
-    def resource_url
-      case resource_type
-      when :static
-        URI.parse source_node.at('StaticResource').content.strip
-      when :iframe
-        URI.parse source_node.at('IFrameResource').content.strip
-      end
-    end
-    
-    # Returns HTML text for html resource
-    def resource_html
-      if resource_type == :html
-        source_node.at('HTMLResource').content
-      end
-    end
-    
   end
 end
